add tests

+++ src/main/java/org/example/Calculator.java
@@ -9,7 +9,7 @@ public class Calculator {
         return a - b;
     }
 
-    public float division(float a, float b) {
+    public int division(int a, int b) {
         return a / b;
     }
